requirements.txt
fastapi
uvicorn
requests
python-dotenv
transformers
pymongo
rasa
langchain